hash, database: refactor database logic to use specialized hashSets
Create a custom datastructure hashSet that is optimized for storing
sets of hashes that are most likely to be 4 bytes in length.
Use this data structure instead of naive maps to reduce memory
inefficiencies resulting from storing string structs to memories on
the heap.

Furthermore, flip the dimensionality of the database lookup structure.
Rather than being a map (keyed by hashes) to slices of ThreatDescriptors,
make it so that it is a map (keyed by ThreatDescriptors) to hashSets.
This is more efficient since that are only expected to be a small handful
of ThreatDescriptors, while we can reasonably assume that there will
be millions of hashes.

Lastly, for the database update structure, avoid storing a full list
of all the hashes. We can already regenerate these from the hashSets
when it comes time to actually update (which is very rare from the
perspective of a CPU and memory system). Thus, there is no point
holding on to memory that can be regenerated from the hashSet.