database: add some timing jitter for staleness detection
Let TP be the update period (the time between calls to the list update API).
Let TU be the amount of time that the list update API takes.

If we start the update every TP, then there is a TU window of time where we
will consider the database to be stale since the update is still going on.
Thus, we should give some leeway to the staleness calculation to account for
the update itself.