database: refactor load and save to be standalone functions (#26)
The state of the database object was heavily dependent on the current
stage that the update process was currently in. This makes it difficult
doing a save since it's not obvious what state is stable and what is
in fluctuation.

Instead of load and save being methods of database, make them standalone
functions where the caller is responsible for explicitly passing in what
is to be saved.