Find a flaw in parsing URL
Function getScheme in urls.go regards the string before the first symbol
':' as protocol and the path is after ':'. So sbserver  will report
"invalid path" if the url's format is "domain:port/path". But I think it
should be "http://domain/path" after being canonicalized.
So I add a function to parse the string after ':'. It will return ("",
url) if it is a port, so the protocol is empty and the rest is
"domain:port/path". And (url[:i], url[i+1:]) will be returned in other
stations.
For example, the result should be "http://www.gotaport.com/" after
"http://www.gotaport.com:1234/" and "www.gotaport.com:1234/" being
canonicalized. But sbserver's results are "http://www.gotaport.com/" and
"invalid path".