vendor: fix import path of protobuf
A prior commit to add vendor accidentally used the wrong import path
for the protobuf package. We fix this path:
	github.com/protobuf -> github.com/golang/protobuf

We also commit the protoc-gen-go package into protobuf since it is a
dependency for the unit tests. I left that out in the original commit
for vendoring to keep the dependency tree smaller.

We also add go1.7 to the list of Go versions to test for.
We remove tip from the list since it can be fairly noisy.