Fixes panic on nil cache duration.
Prior to this commit if a (misbehaving/buggy) server returns
a `FindFullHashesRequest` that includes one or more `ThreatMatch`
objects that have a nil `CacheDuration` field the `cache.Update()` call
will panic.

This commit uses a default duration of `time.Second` when there is no
`CacheDuration` set to avoid the nil panic.