Adds `-server` flag to `sblookup`.
This commit adds a new `-server` flag to the `sblookup` utility to allow
specifying the API server to use for lookups. This is convenient for
testing against the provided `sbserver` proxy, or a non-standard (e.g.
mocked) GSB server. If it isn't provided the value defaults to
`safebrowsing.DefaultServerURL`, preserving the `sblookup` behaviour
prior to this commit.