`nil` panic fixes for misbehaving/buggy servers. (#45)
* Fixes panic on nil checksum.

Prior to this commit if a (misbehaving/buggy) server returns
a `FetchThreatListUpdatesResponse` that has one or more `FetchThreatListUpdatesResponse_ListUpdateResponse` objects with a nil
`Checksum` field the `threatsForUpdate.update()` call will panic.

This commit guards the assignment of `phs.SHA256` based on whether the
field is nil or not. If the field is nil we skip processing `phs.SHA256`
and the `bytes.Equal` check will fail as expected without producing
a panic.

* Fixes panic on nil cache duration.

Prior to this commit if a (misbehaving/buggy) server returns
a `FindFullHashesRequest` that includes one or more `ThreatMatch`
objects that have a nil `CacheDuration` field the `cache.Update()` call
will panic.

This commit uses a default duration of `0s` when there is no
`CacheDuration` set to avoid the nil panic.