urls.go fix for Go 1.8
Go 1.8 changed the behavior of url.Parse to reject urls like
"this_that:other/thing" in accordance with RFC 3986. This also rejects
an IPv6 address without a proper schema. Since this is something we want
to support, we instead use the host already parsed from canonicalHost()
as an input to net.ParseIP().