Respect MinimumWaitDuration and do exponential failure backoff
Modifies how the database Update function behaves by having it return
a delay that changes the amount of time between updates. By default
this will either be MinimumWaitDuration or the configured UpdatePeriod
plus a random jitter.

If the update api fails, then it will now keep track of the number of
times it fails and do an exponential failure backoff. The initial retry
is around 15 minutes after the failure and it goes up from there.