Fix time.Time comparisons for upcoming Go1.9 monotonic times
An upcoming change in Go1.9 changes the time.Time type to make a measurement
of both the wall clock and a monotonic clock. The existence of a monotonic
measurement makes it such that direct equality and use of reflect.DeepEqual
now fail. Fix those use cases by stripping out the monotonic measurement
so that equality continues to work.