Implement WaitUntilReady(ctx) (#61)
* Implement WaitUntilReady(ctx)

Allows callers to block until the database is in a non-error state, or
until the provided context is cancelled.

Also cleaned up some of the errors set in Init that marked the DB as
stale when it was actually just completely broken.

This is the second part of issue #44

* Make WaitUntilReady respect closed SafeBrowsers.

Moves the 'selection' to the safebrowser to be able to inspect the
open/closed state of the instance. The database now has a Ready()
method instead, which just returns the current readyCh, similar to
context's Done() method.