Increase grace period for stale database content.
Prior to this commit the database was considered stale when the time
since the last update was greater than the update period plus jitter.
Since the update to refresh the content **also** wasn't scheduled until
update period plus jitter time had elapsed there was a period of overlap
in which all lookups fail since the database is marked stale but the
update hasn't completed yet.

This commit increases the grace period for a stale database and only
marks the database inconsistent when it has been longer than **two**
times the update period plus jitter. This allows for an update request
to fail without causing the database to be immediately inconsistent and
removes the small window of inconsistency when the update succeeds.

A new `isStale` function is added to the database to make the
calculation of staleness against a given last update time based on the
current time. Both `Init` and `Status` use `isStale` to determine
staleness of a particular database state.

This commit adds a new test vector to `TestDatabaseInit` that loads
a old but not yet stale (update period + 30m old) database. This commit
additionally increases the age of the existing stale database test
vector to be 2*update period + jitter. Lastly this commit adds
a `TestIsStale` unit test that tests the `isStale` function of the
database.