Exposes a function to check if a string is a URL.
This makes it possible to check if a string will pass URL parsing before
calling LookupURLs. This can help identify if one string in a batch will
cause an error that will abort the LookupURLs call.

A pitfall of this method is that any series of words without punctuation
is considered a valid URL, as the code will turn spaces into '%20's and
consider the whole string as a long hostname.

Fixes #72